refactor(mcp): extract runtime catalog tools
